cleanup after implementing autogeneration of IDs

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/AccountRoleRepository.java
@@ -5,5 +5,5 @@ import ca.mcgill.ecse321.grocerystore.model.AccountRole;
 
 public interface AccountRoleRepository extends CrudRepository<AccountRole, Integer> {
 	
-	AccountRole findByRoleID(Integer roleID);
+	AccountRole findByRoleID(Long roleID);
 }
