+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/OwnerRepository.java
@@ -7,5 +7,5 @@ import ca.mcgill.ecse321.grocerystore.model.Owner;
 
 public interface OwnerRepository extends CrudRepository<Owner, Integer> {
 	
-	Owner findByRoleID(Integer roleID);
+	Owner findByRoleID(Long roleID);
 }
